Test suite update and BF for Agilent 34410a
Test suite extended for full coverage.

Minor bug fix: Returning `int('9.91000000E+37')` in `read_last_data`
returns a `ValueError`, and is not consistent with the routine otherwise
returning a `float`. Changed to returning a float.

+++ instruments/agilent/agilent34410a.py
@@ -162,7 +162,7 @@ class Agilent34410a(SCPIMultimeter):  # pylint: disable=abstract-method
         }
 
         if data == '9.91000000E+37':
-            return int(data)
+            return float(data)
         else:
             data = data.split(" ")
             data[0] = float(data[0])
